refactor(sidebar): move account refresh tests to a sibling file
The inline module passed the 500-line limit the invariants gate enforces.
